pyproject.toml
[project]
name = "rag-medical-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "ipykernel>=6.30.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "langchain==0.3.26",
    "langchain-community==0.3.26",
    "langchain-groq==0.3.5",
    "langchain-pinecone==0.2.8",
    "pypdf==5.6.1",
    "python-dotenv==1.1.0",
    "rag-medical-app",
    "sentence-transformers==4.1.0",
]

[tool.uv.sources]
rag-medical-app = { workspace = true }
